requirements.txt
grpcio
grpcio-tools
pytest
pytest-icdiff
requests
